Refactored enemy ai some, more natural movement
Instead of automatically lining up with the player to mathematical
precision, they now use the decision making system to change direction
to the player only when making a decision, so they can over/under shoot
the player, which makes for better ai as they are not "TOO" good.  Also
it makes them seem more human.